demo/requirements.txt
pathway[all]
streamlit==1.31.0
llama-index==0.10.68
python-dotenv==1.0.1
traceloop-sdk==0.12.5
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
llama-index-llms-ollama
llama-index-retrievers-pathway